applied_python/bot/requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.1.0
certifi==2025.1.31
contourpy==1.3.0
cycler==0.12.1
fonttools==4.56.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
magic-filter==1.0.12
matplotlib==3.9.4
multidict==6.1.0
numpy==2.0.2
packaging==24.2
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
yarl==1.18.3
zipp==3.21.0
